backend/requirements.txt
fastapi
uvicorn
youtube-transcript-api
sentence-transformers
faiss-cpu
transformers
torch
